+++ MsSql/Test/quick_clean_test.sql
@@ -2,6 +2,9 @@
 PRINT 'Testing clean command...'
 
 -- Simple test 
-EXEC [dbo].[Zync] 'clean'
+IF OBJECT_ID('[dbo].[Zync]', 'P') IS NOT NULL
+	EXEC [dbo].[Zync] 'clean'
+ELSE
+	PRINT 'Zync procedure not found. Skipping clean.'
 
 PRINT 'Clean command completed.'